reworked FlashableCalliope with seperation of V1/V2 and V3 functionality in first (experimental) form
