+++ Calliope App/ViewsAndControllers/EditorsAndPrograms/Editors/Playground/PlaygroundStartViewController.swift
@@ -21,7 +21,7 @@ class PlaygroundStartViewController: UIViewController {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         self.rearrangeStackview(view.bounds.size)
-        MatrixConnectionViewController.instance?.calliopeClass = FlashableCalliope.self
+        MatrixConnectionViewController.instance?.calliopeClass = DiscoveredBLEDDevice.self
     }
 
     override func viewDidAppear(_ animated: Bool) {
