+++ Calliope App/ViewsAndControllers/Home/Onboarding/TutorialPageConnect.swift
@@ -64,10 +64,10 @@ class TutorialPageConnect: TutorialPageViewController, AnimatingTutorialViewCont
         super.viewWillAppear(animated)
         checkHasConnected()
         readyCalliopeObservation = NotificationCenter.default
-            .addObserver(forName: CalliopeBLEDevice.usageReadyNotificationName, object: nil, queue: OperationQueue.main,
+            .addObserver(forName: DiscoveredBLEDDevice.usageReadyNotificationName, object: nil, queue: OperationQueue.main,
                          using: { _ in self.updateConnectionState() })
         disconnectedCalliopeObservation = NotificationCenter.default
-            .addObserver(forName: CalliopeBLEDevice.disconnectedNotificationName, object: nil, queue: OperationQueue.main,
+            .addObserver(forName: DiscoveredBLEDDevice.disconnectedNotificationName, object: nil, queue: OperationQueue.main,
                          using: { _ in self.updateConnectionState() })
         updateConnectionState()
     }
